refactor(app): rename createErrorLogger to createNetworkErrorLink
