+++ packages/app/src/graphql/createNetworkErrorLink.ts
@@ -1,6 +1,6 @@
 import { onError } from "apollo-link-error";
 import { print } from "graphql/language";
-import createOverlay from "./errorOverlay";
+import createErrorOverlay from "./createErrorOverlay";
 import { boolean } from "boolean";
 
 export default () => {
@@ -8,7 +8,7 @@ export default () => {
         const debug = boolean(process.env.REACT_APP_DEBUG);
 
         if (networkError && debug) {
-            createOverlay({ query: print(operation.query), networkError });
+            createErrorOverlay({ query: print(operation.query), networkError });
         }
     });
 };
