+++ packages/app/src/graphql/index.ts
@@ -1,4 +1,4 @@
 export { default as createOmitTypenameLink } from "./createOmitTypenameLink";
 export { default as createSetContextLink } from "./createSetContextLink";
 export { default as createConsoleLink } from "./createConsoleLink";
-export { default as createErrorLogger } from "./createErrorLogger";
+export { default as createNetworkErrorLink } from "./createNetworkErrorLink";
